Add files via upload
The current program I've been working on, the relevant files are
/NFLscrape/src/main/java/main/NFLscrape/DataBuilderAndHashMap.java and
/NFLscrape/src/main/java/main/NFLscrape/Helpers.java
it is in the end stages of development, I want to organize and remove useless testing files, I want to debug the "vegas line" datapoint getter, I want to implement the coach data scraper and hashmap (however I may also save this for future iterations of this program because it increases the time the scraper takes to finish, it is out of scope of the project for now, maybe if I want improvements to the dataset I'll try it out), I want to implement writing to a csv file after each game is finished being scraped, and potentially I'll implement weekly data hashmap updates for each player in order to make rookie data a little more fleshed out. This scraper when done will be used to create 2 datasets. the first having each row being used as an input to various ANNs that attempt to classify winner/loser for home/away team, total points scored regression, away/home points scored regression. The second dataset will be used to make a 3d graph visualization of the connections between QBs, Receivers, and Defenders for the last 21 years, a connection to a receiver being minimum 10 (or something else like 20 idk yet must look at data) completions to that receiver, and to defender is any interception, forced fumble or sack.