Delete nflDataset2.csv
This data is already present within the ProFootballReferenceScraper directory, just shows how much data is in 1 row, much of it has negligible impact on game outcomes such as score and winner, for example kick and punt returner impact on game outcomes is very low (except the starting field position which is impactful but is not contained within the data on the scraped site). Because of the size of the data (655 cols) it may be harder for the neural network to draw classifications from it, which is why getting rid of irrelevant or negligible impact data may be beneficial. It is also beneficial in the scraping process as the more data it scrapes the more edge cases I have to consider.